Refactor AudioInterface to include a method for determining WAV file bit depth and remove deprecated utility function
