+++ src/neuroconv/datainterfaces/behavior/audio/audiointerface.py
@@ -1,4 +1,5 @@
 import json
+import wave
 from pathlib import Path
 from typing import Literal, Optional
 
@@ -8,7 +9,7 @@ from pydantic import FilePath, validate_call
 from pynwb import NWBFile
 
 from ....basetemporalalignmentinterface import BaseTemporalAlignmentInterface
-from ....tools.audio import add_acoustic_waveform_series, get_wav_bit_depth
+from ....tools.audio import add_acoustic_waveform_series
 from ....utils import (
     get_base_schema,
 )
@@ -211,7 +212,7 @@ class AudioInterface(BaseTemporalAlignmentInterface):
 
         for file_index, (acoustic_waveform_series_metadata, file_path) in enumerate(zip(audio_metadata, file_paths)):
             # Check if the WAV file is 24-bit (3 bytes per sample)
-            bit_depth = get_wav_bit_depth(file_path)
+            bit_depth = self._get_wav_bit_depth(file_path)
             
             # Memory mapping is not compatible with 24-bit WAV files
             use_mmap = bit_depth != 24
@@ -232,3 +233,23 @@ class AudioInterface(BaseTemporalAlignmentInterface):
             )
 
         return nwbfile
+
+    @staticmethod
+    def _get_wav_bit_depth(file_path):
+        """
+        Get the bit depth of a WAV file.
+        
+        Parameters
+        ----------
+        file_path : str or Path
+            Path to the WAV file
+            
+        Returns
+        -------
+        int
+            Bit depth of the WAV file (8, 16, 24, 32, etc.)
+        """
+        with wave.open(str(file_path), 'rb') as wav_file:
+            sample_width = wav_file.getsampwidth()
+            bit_depth = sample_width * 8
+        return bit_depth
