+++ src/neuroconv/tools/audio/__init__.py
@@ -1,2 +1,1 @@
 from .audio import add_acoustic_waveform_series
-from .audio_utils import get_wav_bit_depth
